backend/pyproject.toml
[tool.poetry]
name = "algofinstatix"
version = "0.1.0"
description = "AlgoFinStatiX - Financial Analytics Platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "backend/src" },
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.1.0"
pydantic = "^2.11.4"
pydantic-settings = "^2.9.1"
email-validator = "^2.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
asyncpg = "^0.30.0"
alembic = "^1.16.1"
sqlalchemy-utils = "^0.41.2"
python-dateutil = "^2.9.0"
anyio = "^4.9.0"
httpx = "^0.28.1"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"
ruff = "^0.11.10"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 88
target-version = "py313"
select = ["E", "F", "I"]

[tool.ruff.isort]
known-first-party = ["src"]
